Switch test suite to capybara based one instead of unicorn/net-http/hpricot hybrid
While here, add capybara-webkit based integration test suite with unicorn,
only used on ruby 1.9+.

Fix some small issues found by said test suite.

+++ Rakefile
@@ -5,10 +5,11 @@ task :default do
     ENV['RAILS_ENV'] = 'test'
     sh "echo -n '' > log/rails.test.log"
     sh "echo -n '' > log/unicorn.test.log"
-    sh "#{FileUtils::RUBY} -S unicorn -p 8989 -c config/unicorn.test.conf -D"
+    sh "#{FileUtils::RUBY} -S unicorn -p 8989 -c config/unicorn.test.conf -D" if RUBY_VERSION > '1.9'
     sh "#{FileUtils::RUBY} -S spec unit_test.rb"
     sh "#{FileUtils::RUBY} -S spec test.rb"
+    sh "#{FileUtils::RUBY} -S spec test_ajax.rb" if RUBY_VERSION > '1.9'
   ensure
-    sh "kill `cat log/unicorn.test.pid`"
+    sh "kill `cat log/unicorn.test.pid`" if RUBY_VERSION > '1.9'
   end
 end
